# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OneTap Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get a user's public profile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/achievements:
    get:
      operationId: getUserAchievements
      tags: [achievements]
      summary: Get achievements earned by a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAchievement"

  # ─── FRIENDS ─────────────────────────────────────────────────────────────────
  /friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List current user's friends
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Friends list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friendship"
  /friends/request:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
        "409":
          description: Already friends or request pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /friends/requests:
    get:
      operationId: listIncomingRequests
      tags: [friends]
      summary: List incoming friend requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Incoming requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"
  /friends/requests/sent:
    get:
      operationId: listSentRequests
      tags: [friends]
      summary: List sent friend requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sent requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"
  /friends/requests/{requestId}:
    patch:
      operationId: respondToFriendRequest
      tags: [friends]
      summary: Accept or reject a friend request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /friends/{friendId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      security:
        - bearerAuth: []
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Friendship removed
        "404":
          description: Friend not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── ROOMS ───────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "204":
          description: All marked as read

  # ─── SEASONS ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── AUTH SCHEMAS ─────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, username, email, level, xp, language, isPlusMember, onlineStatus, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]

        bio:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
        country:
          type: ["string", "null"]
        language:
          type: string
          enum: [pt, en]
        isPlusMember:
          type: boolean
        onlineStatus:
          type: string
          enum: [online, away, offline]
        gamesPlayed:
          type: integer
        wins:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        bio:
          type: string
          maxLength: 200
        avatarUrl:
          type: string
        language:
          type: string
          enum: [pt, en]
        country:
          type: string

    # ─── FRIEND SCHEMAS ───────────────────────────────────────────────────────
    FriendRequestInput:
      type: object
      required: [toUserId]
      properties:
        toUserId:
          type: string
    FriendRequestUpdate:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, reject]
    FriendRequest:
      type: object
      required: [id, fromUserId, toUserId, status, createdAt]
      properties:
        id:
          type: string
        fromUserId:
          type: string
        toUserId:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        fromUser:
          $ref: "#/components/schemas/UserProfile"
        toUser:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time
    Friendship:
      type: object
      required: [id, userId, friendId, friend, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        friendId:
          type: string
        friend:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time

    # ─── ROOM SCHEMAS ─────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, type, title, body, isRead, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [friend_request, friend_accepted, room_invite, achievement_unlocked, season_start, season_end, system]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        data:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    # ─── SEASON SCHEMAS ───────────────────────────────────────────────────────
...
    Achievement:
      type: object
      required: [id, key, title, description, icon, category, xpReward]
      properties:
        id:
          type: string
        key:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
          enum: [social, competitive, explorer, milestone, seasonal]
        xpReward:
          type: integer
        isSecret:
          type: boolean
    UserAchievement:
      type: object
      required: [achievement, unlockedAt]
      properties:
        achievement:
          $ref: "#/components/schemas/Achievement"
        unlockedAt:
          type: string
          format: date-time

    # ─── RANKING SCHEMAS ──────────────────────────────────────────────────────
...